Move some tx overview components to common folder
