+++ packages/app/src/features/dialogs/common/components/transaction-overview/TransactionOverviewDetailsItem.tsx
