+++ packages/app/src/features/dialogs/e-mode/components/LTVChange.tsx
@@ -1,9 +1,8 @@
 import { formatPercentage } from '@/domain/common/format'
 import { Percentage } from '@/domain/types/NumericValues'
 import { assets } from '@/ui/assets'
-
 import { testIds } from '@/ui/utils/testIds'
-import { TransactionOverviewDetailsItem } from '../../common/components/TransactionOverviewDetailsItem'
+import { TransactionOverviewDetailsItem } from '../../common/components/transaction-overview/TransactionOverviewDetailsItem'
 
 interface LTVChangeProps {
   currentMaxLTV: Percentage
