+++ packages/app/src/features/dialogs/savings/migrate/common/components/TransactionOverview.tsx
@@ -2,11 +2,13 @@ import { formatPercentage } from '@/domain/common/format'
 import { Token } from '@/domain/types/Token'
 import { DialogPanel } from '@/features/dialogs/common/components/DialogPanel'
 import { DialogPanelTitle } from '@/features/dialogs/common/components/DialogPanelTitle'
-import { TransactionOverviewDetailsItem } from '@/features/dialogs/common/components/TransactionOverviewDetailsItem'
+import { RouteItem } from '@/features/dialogs/common/components/transaction-overview/RouteItem'
+import { SkyBadge } from '@/features/dialogs/common/components/transaction-overview/SkyBadge'
+import { TransactionOverviewDetailsItem } from '@/features/dialogs/common/components/transaction-overview/TransactionOverviewDetailsItem'
 import { assets } from '@/ui/assets'
 import { testIds } from '@/ui/utils/testIds'
 import { assert } from '@/utils/assert'
-import { RouteItem, SkyBadge, TransactionOutcome } from '../../../common/components/transaction-overview/components'
+import { TransactionOutcome } from '../../../common/components/transaction-overview/components'
 import { TransactionOverviewPlaceholder } from '../../../common/components/transaction-overview/components/TransactionOverviewPlaceholder'
 import { MigrateDialogTxOverview } from '../types'
