+++ packages/app/src/features/dialogs/withdraw/components/WithdrawOverviewPanel.tsx
@@ -2,10 +2,9 @@ import { formatPercentage } from '@/domain/common/format'
 import { TokenWithValue } from '@/domain/common/types'
 import { DialogPanel } from '@/features/dialogs/common/components/DialogPanel'
 import { DialogPanelTitle } from '@/features/dialogs/common/components/DialogPanelTitle'
-
-import { HealthFactorChange } from '../../common/components/HealthFactorChange'
-import { TokenValueChange } from '../../common/components/TokenValueChange'
-import { TransactionOverviewDetailsItem } from '../../common/components/TransactionOverviewDetailsItem'
+import { HealthFactorChange } from '../../common/components/transaction-overview/HealthFactorChange'
+import { TokenValueChange } from '../../common/components/transaction-overview/TokenValueChange'
+import { TransactionOverviewDetailsItem } from '../../common/components/transaction-overview/TransactionOverviewDetailsItem'
 import { PositionOverview } from '../logic/types'
 
 export interface WithdrawOverviewPanelProps {
